management-node/management/requirements.txt
Flask==2.0.1
Flask-JWT-Extended==4.3.1
mysql-connector-python==8.0.31
Werkzeug==2.0.3